+++ app/javascript/src/application/item_show.js
@@ -4,12 +4,13 @@ document.addEventListener('turbolinks:load', () => {
     $('.js-download').multiDownload();
   });
 
-  $('#more-information-hidden').on('hidden.bs.collapse', () => {
-    $('.js-shown-text').toggleClass('d-none');
-    $('.js-hidden-text').toggleClass('d-none');
+  $('#more-information-hidden').on('hidden.bs.collapse shown.bs.collapse', () => {
+    $('.js-more-information-shown').toggleClass('d-none');
+    $('.js-more-information-hidden').toggleClass('d-none');
   });
-  $('#more-information-hidden').on('shown.bs.collapse', () => {
-    $('.js-shown-text').toggleClass('d-none');
-    $('.js-hidden-text').toggleClass('d-none');
+
+  $('#edit-history-hidden').on('hidden.bs.collapse shown.bs.collapse', () => {
+    $('.js-edit-history-shown').toggleClass('d-none');
+    $('.js-edit-history-hidden').toggleClass('d-none');
   });
 });