+++ test/application_system_test_case.rb
@@ -3,6 +3,8 @@ require 'selenium-webdriver'
 
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
 
+  Capybara.default_max_wait_time = 5
+
   if ENV['CAPYBARA_NO_HEADLESS']
     driven_by :selenium, using: :chrome, screen_size: [1400, 1400]
   else
@@ -25,6 +27,20 @@ class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
 
   def logout_user
     visit logout_url
+  end
+
+  # Used to enable papertrail gem (disabled by default in tests).
+  def with_versioning
+    was_enabled = PaperTrail.enabled?
+    was_enabled_for_request = PaperTrail.request.enabled?
+    PaperTrail.enabled = true
+    PaperTrail.request.enabled = true
+    begin
+      yield
+    ensure
+      PaperTrail.enabled = was_enabled
+      PaperTrail.request.enabled = was_enabled_for_request
+    end
   end
 
 end