+++ test/controllers/items/draft_controller_test.rb
@@ -2,18 +2,9 @@ require 'test_helper'
 
 class Items::DraftControllerTest < ActionDispatch::IntegrationTest
 
-  def before_all
-    super
-    @community = Community.create!(title: 'Books', description: 'a bunch of books', owner_id: users(:admin).id)
-    @collection = Collection.create!(title: 'Fantasy Books',
-                                     description: 'some fantasy books',
-                                     owner_id: users(:admin).id,
-                                     community_id: @community.id)
-  end
-
   setup do
-    # some of the cleanup in integration tests is sloppy, so we need to manually clean up Solr
-    JupiterCore::SolrServices::Client.instance.truncate_index
+    @community = communities(:books)
+    @collection = collections(:fantasy_books)
     @user = users(:regular)
   end