+++ test/integration/community_show_test.rb
@@ -3,9 +3,7 @@ require 'test_helper'
 # TODO: This test would be better as an System Acceptance test instead?
 class CommunityShowTest < ActionDispatch::IntegrationTest
 
-  def before_all
-    super
-
+  setup do
     # A community with two collections and a logo
     @community1 = Community.create!(title: 'Two collection community', owner_id: users(:admin).id)
     @collection1 = Collection.create!(community_id: @community1.id,
@@ -18,7 +16,7 @@ class CommunityShowTest < ActionDispatch::IntegrationTest
                             filename: 'image-sample.jpeg', content_type: 'image/jpeg'
 
     # A community with no collections
-    @community2 = Community.create!(title: 'Empty community', owner_id: users(:admin).id)
+    @community2 = communities(:community_with_no_collections)
   end
 
   test 'visiting the show page for a community with two collections as an admin' do