+++ test/system/communities_pagination_and_sort_test.rb
@@ -2,7 +2,7 @@ require 'application_system_test_case'
 
 class CommunitiesPaginationAndSortTest < ApplicationSystemTestCase
 
-  def setup
+  setup do
     # for some runs/seeds (like SEED=1099), stale Communities are left over from other tests.
     # We need to assume the communities created here are the only ones!
     Community.delete_all